Make more work for myself
Imports are not currently working, but they are mostly implemented.
I've developed yet another pervasive concept--the CratePath class
is used to represent a path to a rust module. It can be joined with
a RustType to produce absolute or relative paths to types or functions.
This is very cool, but in order for it to work in the code, it must
be used anywhere that I currently take a String to represent a path.