Attempt to correctly generate more models
This commit fixes a bug in ModelFile::registerModel, which caused
5933 compiler errors. Additionally, it introduces major source
reorganization to group ModelContext related logic into separate
packages. Add a TupleContext for those models which are essentially
just type aliases. Currently, there are 237 compiler errors, because
the changes to FileFactory resulted in a few types which are not
currently matched to an IModelContextFactory, so they are missing
from the source tree.