examples/simple: Add SessionCollection and fix Seuss implementation
This adds some extra required components to the "simple" example.
It's no longer simple, but once I've developed a better high-level
API for seuss, I can pull a lot of the boilerplate out. Rewrite the
seuss implementation of the SessionCollection, and remove the
implementation of the SessionService, since it added no value. Add
instructions to test/README for setting up a Python venv to run
the Redfish protocol validator.