generator/axum: Keep track of allowed methods for GET,HEAD requests
The specification requires that an Allow header be present in all
GET and HEAD responses. This commit attaches middleware to all
components in the into_router() method that attaches an Allow header
to any GET/HEAD responses containing the methods that are actually
attached to the route, if the header is not actually present.