Add OpenapiSpecification
This class is responsible for parsing all of the OpenAPI documents
that make up the Redfish Data Model. Previously, we would only
parse the root-level document, openapi.yaml. However, since many
schemas are not included in the root document, we need to parse all
documents in the package in order to get models like Event. This
class is responsible for facilitating the parsing of each individual
document and aggregating the models into a single collection for
generation.