Refactor data model parsing
Previously, we would attempt to load all of the schema files in the
data model directory, and only add a model if it hadn't been
previously discovered. That led to generating a lot of models that
add no value to have--for example, all of the previous versions of
the Resource models, which were identical anyways. Now, we only
load the highest version of the versioned models, and only the
unique unversioned models. See OpenapiSpecification.java for a
clearer explanation.