generator: Define additional attributes on models
This commit introduces IRustExpression, a generic interface for
valid rust expressions that can be rendered in templates. We use
this with the new AdditionalModelAttributesPolicy and CfgAttrExpression
classes to add a #[cfg_attr] expression that optionally derives
valuable::Valuable on the "valuable" crate feature, and removes this
trait from the registry items.